changed structures of models a bit
