+++ src/curvefit/models/gaussian_mixtures.py
@@ -1,9 +1,11 @@
 import numpy as np
 from curvefit.core.functions import gaussian_pdf
+from curvefit.models.base import Model, DataInputs
 
-class GaussianMixtures:
+class GaussianMixtures(Model):
 
     def __init__(self, stride, size, params=None):
+        super().__init__()
         self.params = params
         self.stride = stride 
         self.size = size
@@ -25,23 +27,22 @@ class GaussianMixtures:
             X.append(gaussian_pdf(t, [self.params[0], beta, self.params[2]]))
         X = np.asarray(X).T
         assert X.shape == (len(t), self.size)
-        return X
+        return X, betas
 
-    def _objective_and_gradient(self, x, data):
-        df = data[0]
-        data_specs = data[1]
-        obs = df[data_specs.col_obs]
-        obs_se = df[data_specs.col_obs_se]
-        t = df[data_specs.col_t]
-        self.matrix = self.compute_design_matrix(t)
+    def _objective_and_gradient(self, x, t, obs, obs_se):
+        self.matrix = self.compute_design_matrix(t)[0]
         residuals = (obs - np.dot(self.matrix, x)) / obs_se
         return 0.5 * np.sum(residuals**2), -(self.matrix.T / obs_se).dot(residuals)
 
     def objective(self, x, data):
-        return self._objective_and_gradient(x, data)[0]
+        if self.data_inputs is None:
+            self.data_inputs = self.convert_inputs(data)
+        return self._objective_and_gradient(x, self.data_inputs.t, self.data_inputs.obs, self.data_inputs.obs_se)[0]
 
     def gradient(self, x, data):
-        return self._objective_and_gradient(x, data)[1]
+        if self.data_inputs is None:
+            self.data_inputs = self.convert_inputs(data)
+        return self._objective_and_gradient(x, self.data_inputs.t, self.data_inputs.obs, self.data_inputs.obs_se)[1]
 
     @property
     def bounds(self):
@@ -54,5 +55,18 @@ class GaussianMixtures:
         return x
 
     def predict(self, x, t):
-        matrix = self.compute_design_matrix(t)
+        matrix = self.compute_design_matrix(t)[0]
         return np.dot(matrix, x)
+
+    def convert_inputs(self, data):
+        if isinstance(data, DataInputs):
+            return data
+        
+        df = data[0]
+        data_specs = data[1]
+
+        t = df[data_specs.col_t].to_numpy()
+        obs = df[data_specs.col_obs].to_numpy()
+        obs_se = df[data_specs.col_obs_se].to_numpy()
+        
+        return DataInputs(t=t, obs=obs, obs_se=obs_se)
